# docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AquaTrackAPI",
    "description": "The world's best API for water metering\n\nSite - [https://aqua-track-front-group-five.vercel.app/](https://aqua-track-front-group-five.vercel.app/)",
    "version": "1.0.0",
    "contact": {}
  },
  "servers": [
    {
      "url": "https://project-aqt-api-testing.onrender.com"
    },
    {
      "url": "http://localhost:3000"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/users/reset-password": {
      "post": {
        "tags": ["users"],
        "summary": "reset-password",
        "description": "reset-password",
        "operationId": "resetPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string",
                    "example": "password"
                  },
                  "token": {
                    "type": "string",
                    "example": "reset password token"
                  }
                }
              },
              "examples": {
                "reset-password": {
                  "value": {
                    "password": "password",
                    "token": "reset password token"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "reset-password",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {}
                    },
                    "message": {
                      "type": "string",
                      "example": "Password was successfully reset!"
                    },
                    "status": {
                      "type": "number",
                      "example": 200
                    }
                  }
                },
                "examples": {
                  "reset-password": {
                    "value": {
                      "data": {},
                      "message": "Password was successfully reset!",
                      "status": 200
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "reset-password",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "jwt expired"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "UnauthorizedError"
                    },
                    "status": {
                      "type": "number",
                      "example": 401
                    }
                  }
                },
                "examples": {
                  "reset-password": {
                    "value": {
                      "data": {
                        "message": "jwt expired"
                      },
                      "message": "UnauthorizedError",
                      "status": 401
                    }
                  }
                }
              }
            }
          }
        },
        "security": [{}]
      }
    },
...
    "/users/userinfo": {
      "get": {
        "tags": ["users"],
        "summary": "userinfo",
        "description": "userinfo",
        "operationId": "userinfo",
        "responses": {
          "204": {
            "description": "deleteWater",
            "content": {}
          },
          "400": {
            "description": "deleteWater",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Bad Request.Invalid parameter."
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "BadRequestError"
                    },
                    "status": {
                      "type": "number",
                      "example": 400
                    }
                  }
                },
                "examples": {
                  "deleteWater": {
                    "value": {
                      "data": {
                        "message": "Bad Request.Invalid parameter."
                      },
                      "message": "BadRequestError",
                      "status": 400
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "deleteWater",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Card not found"
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "NotFoundError"
                    },
                    "status": {
                      "type": "number",
                      "example": 404
                    }
                  }
                },
                "examples": {
                  "deleteWater": {
                    "value": {
                      "data": {
                        "message": "Card not found"
                      },
                      "message": "NotFoundError",
                      "status": 404
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["users"],
        "summary": "userinfo",
        "description": "userinfo",
        "operationId": "userinfo1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "type": "string",
                    "example": "2025-03-08T10:20"
                  },
                  "value": {
                    "type": "number",
                    "example": 500
                  }
                }
              },
              "examples": {
                "updateWater": {
                  "value": {
                    "date": "2025-03-08T10:20",
                    "value": 500
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updateWater",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string",
                          "example": "67d03e958f39947802e7ba34"
                        },
                        "created_at": {
                          "type": "string",
                          "example": "2025-03-11T13:45:57.761Z"
                        },
                        "date": {
                          "type": "string",
                          "example": "2025-03-08T10:20"
                        },
                        "updated_at": {
                          "type": "string",
                          "example": "2025-03-11T13:47:41.200Z"
                        },
                        "userId": {
                          "type": "string",
                          "example": "67cabb8b686216dc35c58a85"
                        },
                        "value": {
                          "type": "number",
                          "example": 500
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Sucsesfully updated amount water"
                    },
                    "status": {
                      "type": "number",
                      "example": 200
                    }
                  }
                },
                "examples": {
                  "updateWater": {
                    "value": {
                      "data": {
                        "_id": "67d03e958f39947802e7ba34",
                        "created_at": "2025-03-11T13:45:57.761Z",
                        "date": "2025-03-08T10:20",
                        "updated_at": "2025-03-11T13:47:41.200Z",
                        "userId": "67cabb8b686216dc35c58a85",
                        "value": 500
                      },
                      "message": "Sucsesfully updated amount water",
                      "status": 200
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "updateWater",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Bad Request.Invalid parameter."
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "BadRequestError"
                    },
                    "status": {
                      "type": "number",
                      "example": 400
                    }
                  }
                },
                "examples": {
                  "updateWater": {
                    "value": {
                      "data": {
                        "message": "Bad Request.Invalid parameter."
                      },
                      "message": "Successfully patched a user!",
                      "status": 200
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}